backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cmdstanpy==1.3.0
colorama==0.4.6
contourpy==1.3.2
cryptography==46.0.5
cycler==0.12.1
deprecation==2.1.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
faiss-cpu==1.13.2
fastapi==0.135.1
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.2.0
greenlet==3.3.2
h11==0.16.0
h2==4.3.0
hf-xet==1.4.2
holidays==0.93
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.1
hyperframe==6.1.0
idna==3.11
importlib_resources==6.5.2
joblib==1.5.3
kiwisolver==1.5.0
markdown-it-py==4.0.0
matplotlib==3.10.8
mdurl==0.1.2
mmh3==5.2.1
multidict==6.7.1
nltk==3.9.4
numpy==2.2.6
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
pillow==12.1.1
postgrest==2.28.2
propcache==0.4.1
prophet==1.3.0
psycopg2-binary==2.9.11
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.11.1
PyJWT==2.12.1
pyparsing==3.3.2
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
realtime==2.28.2
requests==2.32.5
rich==14.3.3
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.48
stanio==0.5.1
starlette==0.52.1
storage3==2.28.2
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.2
supabase-auth==2.28.2
supabase-functions==2.28.2
tenacity==9.1.4
threadpoolctl==3.6.0
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.42.0
websockets==15.0.1
yarl==1.23.0
zstandard==0.25.0
